fix flat direction identification for seed points
This commit adds functionality to identify flat directions already at
the step of the Minimize_gen_all function. The deepest minimum is only
selected if its potential value is significantly different and in case
not if the minimum is significantly away a flat direction is identified
and stored. It is ensured that flat directions are then set to zero (or
for multi-dim flat directions a 1D projection is chosen) by updating the
    `ConvertToNonFlatDirections` function.