fix flat direction identification, correct Debyes
This commit refines the identification of the deepest minimum in the seed
point search: Now, if the potential values are close, but the minimum
position are also close enough, still the newly found minimum is
identified as the deepest minimum. Only if the potential values are
close and the minimum positions differ significantly, the old
potentially-deepest-minimum index is kept and we look for a potential
flat direction. This fixes the case of small potential barriers at the
beginning of new phases.

Additionally, this commit sets the dim-6 Debye mass correction in the
RxSMEFT for the scalar singlet direction to zero, as here the scalar singlet is the
external particle and because the EFT term comes with a derivative,
(del_mu S) -> (external momentum * S) = 0.